Issue #2: * Remove old type model based schema generation * Add minimal test for the adapters * small code changes based on suggestions
