--- src/main/java/org/observertc/webrtc/connector/datawarehouses/bigquery/version02/Config.java
@@ -1,4 +1,0 @@
-package org.observertc.webrtc.connector.datawarehouses.bigquery.version02;
-
-public class Config {
-}
