--- src/main/java/org/observertc/webrtc/connector/evaluators/Mapper.java
@@ -1,4 +1,0 @@
-package org.observertc.webrtc.connector.evaluators;
-
-public class Mapper {
-}
