--- src/main/java/org/observertc/webrtc/connector/sinks/bigquery/models/Entry.java
@@ -1,10 +1,0 @@
-package org.observertc.webrtc.connector.sinks.bigquery.models;
-
-import java.util.Map;
-
-public interface Entry {
-
-     EntryType getEntryType();
-
-     Map<String, Object> toMap();
-}
