+++ src/main/java/org/observertc/webrtc/connector/transformations/Mapper.java
@@ -1,0 +1,4 @@
+package org.observertc.webrtc.connector.transformations;
+
+public class Mapper {
+}
