+++ src/test/java/org/observertc/webrtc/connector/transformations/ICECandidatePairMapperTest.java
@@ -1,4 +1,4 @@
-package org.observertc.webrtc.connector.evaluators;
+package org.observertc.webrtc.connector.transformations;
 
 import org.junit.jupiter.api.Test;
